rm nested internals folder
