+++ packages/core/src/http/index.ts
@@ -1,6 +1,5 @@
 export * from './getHTTPStatusCode';
 export * from './resolveHTTPResponse';
-export * from './internals/types';
 export * from './types';
 export * from './batchStreamFormatter';
 export * from './contentType';
