--- packages/core/src/http/internals/types.ts
@@ -1,40 +1,0 @@
-import {
-  AnyRouter,
-  inferRouterContext,
-  inferRouterError,
-  ProcedureType,
-} from '../../';
-import { TRPCError } from '../../error/TRPCError';
-import { TRPCResponse } from '../../rpc';
-import { Dict } from '../../types';
-import { ResponseMeta } from '../types';
-
-export type HTTPHeaders = Dict<string[] | string>;
-
-export interface HTTPResponse {
-  status: number;
-  headers?: HTTPHeaders;
-  body?: string;
-}
-
-export type ResponseChunk = [procedureIndex: number, responseBody: string];
-
-/**
- * @internal
- */
-export type ResponseMetaFn<TRouter extends AnyRouter> = (opts: {
-  data: TRPCResponse<unknown, inferRouterError<TRouter>>[];
-  ctx?: inferRouterContext<TRouter>;
-  /**
-   * The different tRPC paths requested
-   **/
-  paths?: string[];
-  type: ProcedureType | 'unknown';
-  errors: TRPCError[];
-  /**
-   * `true` if the `ResponseMeta` are being
-   * generated without knowing the response data
-   * (e.g. for streaming requests).
-   */
-  eagerGeneration?: boolean;
-}) => ResponseMeta;
